+++ google/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file documents changes to the `workbench-google` library, including notes on how to upgrade to new versions.
 
+## 0.35
+
+SBT Dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google" % "0.35-TRAVIS-REPLACE-ME"`
+
+* disables polling for service account key creation. This reverts to the 0.33 behavior.
 
 ## 0.34